requirements.txt
watchdog>=4.0
pillow-heif>=0.18
Pillow>=10.0
pystray>=0.19
winotify>=1.1
